MBS-9476: Refactor Google OAuth handling to be usable for other subplugins
