+++ lang/de/local_ai_manager.php
@@ -52,6 +52,7 @@ $string['assignpurposes'] = 'Einsatzzwecke festlegen';
 $string['assignrole'] = 'Rolle zuweisen';
 $string['basicsettings'] = 'Grundeinstellungen';
 $string['basicsettingsdesc'] = 'Grundeinstellungen des AI-Managers konfigurieren';
+$string['cachedef_googleauth'] = 'Cache für Google-OAuth2-Access-Token';
 $string['configure_instance'] = 'KI-Tool-Instanzen konfigurieren';
 $string['configureaitool'] = 'KI-Tool konfigurieren';
 $string['configurepurposes'] = 'Einsatzzwecke konfigurieren';
@@ -81,6 +82,8 @@ $string['error_unavailable_noselection'] = 'Dieses Tool ist nur verfügbar, wenn
 $string['error_unavailable_selection'] = 'Dieses Tool ist nur verfügbar, wenn Text markiert wurde.';
 $string['error_userlocked'] = 'Ihr Nutzer wurde von Ihrem Tenant-Manager gesperrt.';
 $string['error_usernotconfirmed'] = 'Für die Nutzung müssen die Nutzungsbedingungen akzeptiert werden.';
+$string['error_vertexai_serviceaccountjsonempty'] = 'Sie müssen den Inhalt der JSON-Datei, die Sie beim Anlegen des Service-Accounts in Ihrer Google-Cloud-Console heruntergeladen haben, einfügen.';
+$string['error_vertexai_serviceaccountjsoninvalid'] = 'Ungültiges Format. Es fehlt der Eintrag "{$a}".';
 $string['exception_curl'] = 'Es ist ein Fehler bei der Verbindung zum externen KI-Tool aufgetreten.';
 $string['exception_curl28'] = 'Die API hat zu lange gebraucht, um Ihre Anfrage zu verarbeiten, oder konnte nicht in angemessener Zeit erreicht werden.';
 $string['exception_default'] = 'Ein allgemeiner Fehler ist aufgetreten, während versucht wurde, die Anfrage an das KI-Tool zu senden.';
@@ -152,6 +155,7 @@ $string['role_extended'] = 'Erweiterte Rolle';
 $string['role_unlimited'] = 'Unbeschränkte Rolle';
 $string['select_tool_for_purpose'] = 'Einsatzzweck "{$a}"';
 $string['selecteduserscount'] = '{$a} ausgewählt';
+$string['serviceaccountjson'] = 'Inhalt der JSON-Datei des Google-Serviceaccounts';
 $string['settingsgeneral'] = 'Allgemein';
 $string['small'] = 'klein';
 $string['squared'] = 'quadratisch';
