+++ lang/en/local_ai_manager.php
@@ -52,6 +52,7 @@ $string['assignpurposes'] = 'Assign purposes';
 $string['assignrole'] = 'Assign role';
 $string['basicsettings'] = 'Basic settings';
 $string['basicsettingsdesc'] = 'Configure basic settings for the AI manager plugin';
+$string['cachedef_googleauth'] = 'Cache for Google OAuth2 access token';
 $string['configure_instance'] = 'Configure AI Tool Instances';
 $string['configureaitool'] = 'Configure AI tool';
 $string['configurepurposes'] = 'Configure the purposes';
@@ -81,6 +82,8 @@ $string['error_unavailable_noselection'] = 'This tool is only available if text 
 $string['error_unavailable_selection'] = 'This tool is only available if no text has been selected.';
 $string['error_userlocked'] = 'Your user has been locked by your tenant manager.';
 $string['error_usernotconfirmed'] = 'You have not accepted the terms of use yet.';
+$string['error_vertexai_serviceaccountjsonempty'] = 'You need to paste the content of the JSON file that you downloaded when creating the service account in your Google Cloud Console.';
+$string['error_vertexai_serviceaccountjsoninvalid'] = 'Invalid format. The entry "{$a}" is missing.';
 $string['exception_curl'] = 'A connection error to the external API endpoint has occurred';
 $string['exception_curl28'] = 'The API took too long to process your request or could not be reached in a reasonable time';
 $string['exception_default'] = 'A general error occurred while trying to send the request to the AI tool';
@@ -152,6 +155,7 @@ $string['role_extended'] = 'extended role';
 $string['role_unlimited'] = 'unlimited role';
 $string['select_tool_for_purpose'] = 'Purpose {$a}';
 $string['selecteduserscount'] = '{$a} selected';
+$string['serviceaccountjson'] = 'Content of the JSON file of the Google service account';
 $string['small'] = 'small';
 $string['squared'] = 'squared';
 $string['statistics_for'] = 'Statistic for {$a}';
