+++ tools/imagen/lang/de/aitool_imagen.php
@@ -24,11 +24,7 @@
  */
 
 $string['adddescription'] = 'Imagen ist ein KI-Modell von Google, das darauf spezialisiert ist, aus Textbeschreibungen Bilder zu generieren.';
-$string['cachedef_imagenauth'] = 'Cache zum Speichern von Authentifizierungstokens';
 $string['err_contentpolicyviolation'] = 'Ihre Anfrage wurde vom Sicherheitssystem des KI-Tools zurückgewiesen. Ihr Prompt enthält vermutlich eine Anweisung, die nicht erlaubt ist.';
 $string['err_predictionmissing'] = 'Auf Ihre Anfrage hin konnte kein Bild generiert werden. Ihr Prompt enthält vermutlich eine Anweisung, die nicht erlaubt ist.';
-$string['err_serviceaccountjsonempty'] = 'Sie müssen den Inhalt der JSON-Datei, die Sie beim Anlegen des Service-Accounts in Ihrer Google-Cloud-Console heruntergeladen haben, einfügen.';
-$string['err_serviceaccountjsoninvalid'] = 'Ungültiges Format. Es fehlt der Eintrag "{$a}".';
 $string['pluginname'] = 'Imagen';
 $string['privacy:metadata'] = 'Das Subplugin "Imagen" des AI-Manager-Plugins speichert keine personenbezogenen Daten.';
-$string['serviceaccountjson'] = 'Inhalt der JSON-Datei des Google-Serviceaccounts';
